Add stand-in type definitions for es2018.asynciterable
Async iterators were standardized as part of the ES2018 specification.
Unfortunately for TypeScript users of this library, this typically means
adding `es2018.asynciterable` to `.compilerOptions.libs` in your
`tsconfig.json`.

This is a bad experience for users. Since that library contains only
type definitions for async iterators, we define them in our project here
to avoid asking users to do this.