Allow queries to be invoked multiple times
Many queries return incorrect results when evaluated twice. Usually, on
second evaluation, it appears that the query has run on an empty
sequence. Methods like `toArray`, `count`, and so on, will return the
corresponding "zero" value (empty array, number 0, etc.). This is not
the way LINQ does it, and it's not the way we should do it.

This occurs because many core JS `Iterable` implementations, when asked
to provide an `Iterator`, do not return "new" `Iterator`. Instead, they
return a singleton. Once completely iterated, there is no built-in way
to restart it.

This commit allows us to evaluate queries twice by removing our
dependency on APIs that return singleton iterators. To help protect
people who accidentally pass `Iterator`s to us (and hence, prevent
themselves from multi-invoke), we change `from` to "wrap" such its input
in an iterator that will re-construct the `Iterable` if it can. This is
particularly helpful when users accidentally pass a `function*`, which
will return the same `Iterator` each time.

This commit will slightly generalize our API to subtly encourage passing
`Iterable` rather than `Iterator` where possible.